fix(server): refactor tile migration to only migrate explicit legacy tile types (#2219)
